+++ CHANGELOG.md
@@ -6,8 +6,7 @@
 - @iov/ethereum: Allow sending of ERC20 tokens
 - @iov/ethereum: Implement `getFeeQuote` method on `EthereumConnection`.
 - @iov/ethereum: Fix transaction fee parsing for transactions from RPC/scraper.
-- @iov/ethereum: Add `blockPollInterval`/`accountPollInterval` to
-  `EthereumConnectionOptions`.
+- @iov/ethereum: Add `pollInterval` to `EthereumConnectionOptions`.
 - @iov/dpos: Let `Serialization.serializeTransaction` accept transactions with
   fee set.